Refactored onboarding
