+++ src/components/Navigation/navigation.js
@@ -1,42 +1,8 @@
-import React, { useCallback } from 'react';
-import { Link, navigate } from '@reach/router';
-// eslint-disable-next-line import/no-unresolved
-import { useDispatch } from 'react-redux';
+import React from 'react';
+import { Link } from '@reach/router';
 import logo from '../../assets/img/logo.png';
-import { auth, getUserInfo } from '../../store/actions';
-import { AuthFlip } from '../Auth/AuthFlip';
-import { useModal } from '../../hooks/useModal';
-
-const login = auth('/auth/login');
-const signup = auth('/auth/register');
 
 export const HorNav = () => {
-  const { isShowing, toggle, close } = useModal();
-
-  const dispatch = useDispatch();
-
-  const loginFn = useCallback(
-    ({ email, password }) =>
-      dispatch(login(email, password))
-        .then(() => {
-          dispatch(getUserInfo()).then(() => {
-            navigate('/dashboard');
-          });
-        })
-        // eslint-disable-next-line no-console
-        .catch(err => console.error(err)),
-    [dispatch]
-  );
-
-  const signupFn = useCallback(
-    ({ email, password, userType: type }) =>
-      dispatch(signup(email, password, type))
-        .then(() => dispatch(getUserInfo()).then(() => navigate('/dashboard')))
-        // eslint-disable-next-line no-console
-        .catch(err => console.error(err)),
-    [dispatch]
-  );
-
   return (
     <>
       <nav className="nav-top">
@@ -57,18 +23,10 @@ export const HorNav = () => {
               Pricing
             </Link>
           </li>
-          <li>
-            <button className="nav-link" type="button" onClick={() => toggle()}>
-              Login
-            </button>
-          </li>
+          <Link className="nav-link" to="/onboarding">
+            Login
+          </Link>
         </ul>
-        <AuthFlip
-          loginSubmit={loginFn}
-          signupFn={signupFn}
-          close={close}
-          isShowing={isShowing}
-        />
       </nav>
     </>
   );
